moved the searchForm and split it into two sepearate forms
